fix(ipc): align IPC implementation with C sources after Phase 4 QA cross-reference
A systematic cross-reference of all Phase 4 code against
`.refs/minix-3.3.0/minix/kernel/` C sources found and fixed 26+
discrepancies:

**IPC core fixes:**
- `will_receive()` was checking the wrong IPC mask direction
- `mini_send()` set RTS_RECEIVING instead of RTS_REPLY_PEND on queued
  senders
- `mini_notify()` used endpoint value instead of privilege slot ID
  (s_id)
  for pending bit storage
- `mini_receive()` failed to clear RTS_RECEIVING before dequeueing
- `do_sync_ipc()` permission check used caller's mask instead of
  caller-to-destination bitmap lookup
- `build_notify_message()` was missing m_notify.timestamp and args
  fields
- `AsynMsg` struct layout verified against C messenger_asyn union

**Grant/safecopy fixes:**
- `verify_grant()` indirect chain now recursively resolves intermediate
  grants instead of stopping at one hop
- `do_safecopy_*` offset arithmetic corrected to grant_start +
  grant_offset
  + caller_offset (was missing the caller offset term)
- `cancel_async()` now walks the async table and sets
  AMF_DONE/AMF_NOTIFY
  per C semantics

**Signal/scheduler fixes:**
- `send_sig()` was missing the SYSTEM notification's source endpoint
  argument
- `cause_sig()` added null-priv guard before signal manager notify
- `notify_scheduler()` message type corrected to SCHEDULING_NO_QUANTUM
- `s_sig_pending` width corrected from u64 to u128 (C sigset_t on
  x86_64)
- `clear_ipc_refs()` cancel_async call is now null-priv guarded

**Constants:**
- `FP_EXISTS` corrected from 0x800 to 0x8000_0000_0000_0000
- `IPC_STATUS_*` shift values verified against C `_IPC_STATUS_*` macros

**Test infrastructure:**
- Replaced stack-local Priv allocations in 4 system tests with a
  static TEST_PRIV_POOL (8 slots, aligned), preventing dangling pointer
  crashes when later tests call clear_ipc_refs/cancel_async

**Documentation:**
- Added comprehensive Phase 4 QA Summary to PORTING_PLAN.md documenting
  every fix, its root cause, and the matching C behavior

+++ crates/kernel/src/priv.rs
@@ -233,7 +233,7 @@ pub struct Priv {
     /// Pending hardware interrupts.
     pub s_int_pending: u64,
     /// Pending signals.
-    pub s_sig_pending: u32,
+    pub s_sig_pending: u128,
 
     /// Synchronous alarm timer.
     pub s_alarm_timer: MinixTimer,
